rename dell.py to storcli.py
